Switch the sign convention of the lags to be consistent with Seabird convention.
RSKTOOLS-4 #comment The lag convention has been inversed to follow Seabird's convention. A negative lag is a forwards shift in time (later) a positive lag in a backwards shift in time (earlier)

RSKTOOLS-29 #comment The lag convention is changed to follow Seabird's. Negative lag is forward shift in time (later), a positive lag is a backwardshift in time (earlier).